Extract version helpers to tests/_helpers.py
Reviewer #1/#2/#3/#5 flagged `from conftest import ...` in
test_url_backend.py as fragile: it relies on pytest's default
--import-mode=prepend adding tests/ to sys.path via the conftest at
that level.  Under --import-mode=importlib the import fails, and
running the test file directly from an unusual cwd is at the mercy of
pytest's rootdir heuristics.

Fix: move URL_BACKEND_FIX_VERSION and the git_annex_version /
git_annex_version_below / git_annex_releases_since helpers into a
plain module tests/_helpers.py.  conftest.py now only holds the
`pytest_report_header` hook (plus its two report-specific helpers
`_first_line` and `_git_annex_summary`), and imports the shared bits
from `_helpers`.

To make `from _helpers import ...` work under any --import-mode,
setup.cfg gets a `[tool:pytest]` section with `pythonpath = tests`.
This is the pytest-native way to guarantee a directory ends up on
sys.path regardless of import-mode / rootdir resolution.

Verified both import modes:

  $ pytest tests                                    # prepend (default)
  4 passed
  $ pytest --import-mode=importlib tests/extra/...  # importlib
  2 passed

Both invocations still print the "extra-tests tools" / git-annex
version / releases-newer-than-installed header.
