+++ setup.cfg
@@ -1,3 +1,9 @@
+[tool:pytest]
+# Put tests/ on sys.path so both the conftest and individual test
+# modules can import shared helpers via `from _helpers import ...`
+# under any --import-mode (prepend, importlib, ...).
+pythonpath = tests
+
 [flake8]
 doctests = True
 #max-doc-length = 100
